requirements.txt
pandas
matplotlib
langchain
sentence-transformers
faiss-cpu
transformers==4.44.2
torch==2.4.1
streamlit